updated abstract methods
